Render real color rects

+++ lib/ecs/src/components/image.rs
@@ -1,6 +1,6 @@
 #[derive(Default)]
 pub struct Image {
   pub source: String,
-  pub width: u32,
-  pub height: u32,
+  pub width: f32,
+  pub height: f32,
 }
